+++ benchmark-base/src/main/java/org/conscrypt/EngineWrapBenchmark.java
@@ -37,6 +37,7 @@ import static org.conscrypt.TestUtils.newTextMessage;
 import static org.junit.Assert.assertEquals;
 
 import java.nio.ByteBuffer;
+import java.util.Locale;
 import javax.net.ssl.SSLEngine;
 import javax.net.ssl.SSLEngineResult;
 import javax.net.ssl.SSLException;
@@ -150,7 +151,8 @@ public final class EngineWrapBenchmark {
         }
         if (result.bytesConsumed() != src.limit()) {
             throw new RuntimeException(
-                    String.format("Operation didn't consume all bytes. Expected %d, consumed %d.",
+                    String.format(Locale.US,
+                            "Operation didn't consume all bytes. Expected %d, consumed %d.",
                             src.limit(), result.bytesConsumed()));
         }
     }
@@ -167,7 +169,7 @@ public final class EngineWrapBenchmark {
 
             @Override
             public EngineType engineType() {
-                return EngineType.CONSCRYPT_POOLED;
+                return EngineType.CONSCRYPT_UNPOOLED;
             }
 
             @Override
