+++ benchmark-caliper/src/main/java/org/conscrypt/CaliperEngineWrapBenchmark.java
@@ -1,0 +1,106 @@
+/*
+ * Copyright 2017 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Copyright 2017 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.conscrypt;
+
+import com.google.caliper.AfterExperiment;
+import com.google.caliper.BeforeExperiment;
+import com.google.caliper.Param;
+import org.conscrypt.EngineWrapBenchmark.Config;
+
+import javax.net.ssl.SSLException;
+
+/**
+ * Benchmark comparing performance of various engine implementations to conscrypt.
+ */
+public class CaliperEngineWrapBenchmark {
+    private final CaliperConfig config = new CaliperConfig();
+
+    @Param({TestUtils.TEST_CIPHER})
+    public String a_cipher;
+
+    @Param
+    public BufferType b_buffer;
+
+    @Param({"64", "512", "4096"})
+    public int c_message;
+
+    @Param
+    public EngineType d_engine;
+
+    private EngineWrapBenchmark benchmark;
+
+    @BeforeExperiment
+    public void setUp() throws Exception {
+        benchmark = new EngineWrapBenchmark(config);
+    }
+
+    @AfterExperiment
+    public void teardown() {
+        benchmark.teardown();
+    }
+
+    public void timeWrap(int reps) throws SSLException {
+        for (int i = 0; i < reps; ++i) {
+            benchmark.wrap();
+        }
+    }
+
+    public void timeWrapAndUnwrap(int reps) throws SSLException {
+        for (int i = 0; i < reps; ++i) {
+            benchmark.wrapAndUnwrap();
+        }
+    }
+
+    private final class CaliperConfig implements Config {
+
+        @Override
+        public BufferType bufferType() {
+            return b_buffer;
+        }
+
+        @Override
+        public EngineType engineType() {
+            return d_engine;
+        }
+
+        @Override
+        public int messageSize() {
+            return c_message;
+        }
+
+        @Override
+        public String cipher() {
+            return a_cipher;
+        }
+    }
+}
