+++ build.gradle
@@ -17,7 +17,10 @@ buildscript {
 }
 
 subprojects {
-    def androidProject = (project.name == 'conscrypt-android') || (project.name == 'conscrypt-android-platform')
+    def androidProject = ((project.name == 'conscrypt-android')
+            || (project.name == 'conscrypt-android-platform')
+            || (project.name == 'conscrypt-benchmark-android')
+            || (project.name == 'conscrypt-benchmark-caliper'))
     if (!androidProject) {
         apply plugin: 'java'
         apply plugin: 'cpp'
