+++ platform/build.gradle
@@ -5,7 +5,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3' // jcenter has the latest
+        classpath 'com.android.tools.build:gradle:2.3.3' // jcenter has the latest
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
     }
 }
