+++ settings.gradle
@@ -3,9 +3,11 @@ include ":conscrypt-android"
 include ":conscrypt-android-platform"
 include ":conscrypt-android-stub"
 include ":conscrypt-api-doclet"
-include ":conscrypt-benchmark-base"
+include ":conscrypt-benchmark-android"
+include ":conscrypt-benchmark-caliper"
 include ":conscrypt-benchmark-graphs"
 include ":conscrypt-benchmark-jmh"
+include ":conscrypt-benchmark-openjdk"
 include ":conscrypt-constants"
 include ":conscrypt-libcore-stub"
 include ":conscrypt-openjdk"
@@ -17,9 +19,11 @@ project(':conscrypt-android').projectDir = "$rootDir/android" as File
 project(':conscrypt-android-platform').projectDir = "$rootDir/platform" as File
 project(':conscrypt-android-stub').projectDir = "$rootDir/android-stub" as File
 project(':conscrypt-api-doclet').projectDir = "$rootDir/api-doclet" as File
-project(':conscrypt-benchmark-base').projectDir = "$rootDir/benchmark-base" as File
+project(':conscrypt-benchmark-android').projectDir = "$rootDir/benchmark-android" as File
+project(':conscrypt-benchmark-caliper').projectDir = "$rootDir/benchmark-caliper" as File
 project(':conscrypt-benchmark-graphs').projectDir = "$rootDir/benchmark-graphs" as File
 project(':conscrypt-benchmark-jmh').projectDir = "$rootDir/benchmark-jmh" as File
+project(':conscrypt-benchmark-openjdk').projectDir = "$rootDir/benchmark-openjdk" as File
 project(':conscrypt-constants').projectDir = "$rootDir/constants" as File
 project(':conscrypt-libcore-stub').projectDir = "$rootDir/libcore-stub" as File
 project(':conscrypt-openjdk').projectDir = "$rootDir/openjdk" as File
